# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

/**
 * Kind of resource
 */
export type ProfessionResourceKind = typeof ProfessionResourceKind[keyof typeof ProfessionResourceKind];


export const ProfessionResourceKind = {
  directory: 'directory',
  prerequisites: 'prerequisites',
} as const;

export interface ProfessionResource {
  label: string;
  /** Kind of resource */
  kind: ProfessionResourceKind;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  url?: string | null;
}

export interface Profession {
  id: number;
  slug: string;
  name: string;
  category: string;
  tagline: string;
  description: string;
  degree: string;
  /** @nullable */
  typicalTimeline?: string | null;
  resources: ProfessionResource[];
}

export type TargetSchoolStatus = typeof TargetSchoolStatus[keyof typeof TargetSchoolStatus];


export const TargetSchoolStatus = {
  researching: 'researching',
  planning: 'planning',
  applying: 'applying',
  applied: 'applied',
  interview: 'interview',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export type TargetSchoolPriority = typeof TargetSchoolPriority[keyof typeof TargetSchoolPriority];


export const TargetSchoolPriority = {
  reach: 'reach',
  target: 'target',
  safety: 'safety',
} as const;

export interface TargetSchool {
  id: number;
  schoolName: string;
  programName: string;
  professionSlug: string;
  /** @nullable */
  state?: string | null;
  inState?: boolean;
  status: TargetSchoolStatus;
  priority: TargetSchoolPriority;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  appPortal?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type TargetSchoolInputStatus = typeof TargetSchoolInputStatus[keyof typeof TargetSchoolInputStatus];


export const TargetSchoolInputStatus = {
  researching: 'researching',
  planning: 'planning',
  applying: 'applying',
  applied: 'applied',
  interview: 'interview',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export type TargetSchoolInputPriority = typeof TargetSchoolInputPriority[keyof typeof TargetSchoolInputPriority];


export const TargetSchoolInputPriority = {
  reach: 'reach',
  target: 'target',
  safety: 'safety',
} as const;

export interface TargetSchoolInput {
  /** @minLength 1 */
  schoolName: string;
  /** @minLength 1 */
  programName: string;
  /** @minLength 1 */
  professionSlug: string;
  state?: string;
  inState?: boolean;
  status?: TargetSchoolInputStatus;
  priority?: TargetSchoolInputPriority;
  deadline?: string;
  appPortal?: string;
  notes?: string;
}

export type TargetSchoolUpdateStatus = typeof TargetSchoolUpdateStatus[keyof typeof TargetSchoolUpdateStatus];


export const TargetSchoolUpdateStatus = {
  researching: 'researching',
  planning: 'planning',
  applying: 'applying',
  applied: 'applied',
  interview: 'interview',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export type TargetSchoolUpdatePriority = typeof TargetSchoolUpdatePriority[keyof typeof TargetSchoolUpdatePriority];


export const TargetSchoolUpdatePriority = {
  reach: 'reach',
  target: 'target',
  safety: 'safety',
} as const;

export interface TargetSchoolUpdate {
  /** @minLength 1 */
  schoolName?: string;
  /** @minLength 1 */
  programName?: string;
  /** @minLength 1 */
  professionSlug?: string;
  /** @nullable */
  state?: string | null;
  inState?: boolean;
  status?: TargetSchoolUpdateStatus;
  priority?: TargetSchoolUpdatePriority;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  appPortal?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type PrereqCourseStatus = typeof PrereqCourseStatus[keyof typeof PrereqCourseStatus];


export const PrereqCourseStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface PrereqCourse {
  id: number;
  professionSlug: string;
  name: string;
  category: string;
  status: PrereqCourseStatus;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  credits?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type PrereqCourseInputStatus = typeof PrereqCourseInputStatus[keyof typeof PrereqCourseInputStatus];


export const PrereqCourseInputStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface PrereqCourseInput {
  /** @minLength 1 */
  professionSlug: string;
  /** @minLength 1 */
  name: string;
  category?: string;
  status?: PrereqCourseInputStatus;
  grade?: string;
  credits?: number;
  notes?: string;
}

export type PrereqCourseUpdateStatus = typeof PrereqCourseUpdateStatus[keyof typeof PrereqCourseUpdateStatus];


export const PrereqCourseUpdateStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  completed: 'completed',
} as const;

export interface PrereqCourseUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  status?: PrereqCourseUpdateStatus;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  credits?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface ProfessionCount {
  professionSlug: string;
  professionName: string;
  count: number;
}

export interface DashboardSummary {
  totalSchools: number;
  totalProfessionsExplored: number;
  prereqTotal: number;
  prereqCompleted: number;
  prereqInProgress: number;
  statusCounts: StatusCount[];
  professionCounts: ProfessionCount[];
}

export type ListTargetSchoolsParams = {
professionSlug?: string;
status?: string;
};

export type ListPrereqCoursesParams = {
professionSlug?: string;
};