# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfessionCount } from './professionCount';
import type { StatusCount } from './statusCount';

export interface DashboardSummary {
  totalSchools: number;
  totalProfessionsExplored: number;
  prereqTotal: number;
  prereqCompleted: number;
  prereqInProgress: number;
  statusCounts: StatusCount[];
  professionCounts: ProfessionCount[];
}