# lib/api-zod/src/generated/types/prereqCourseUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PrereqCourseUpdateStatus } from './prereqCourseUpdateStatus';

export interface PrereqCourseUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  status?: PrereqCourseUpdateStatus;
  /** @nullable */
  grade?: string | null;
  /** @nullable */
  credits?: number | null;
  /** @nullable */
  notes?: string | null;
}