# lib/api-zod/src/generated/types/profession.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfessionResource } from './professionResource';

export interface Profession {
  id: number;
  slug: string;
  name: string;
  category: string;
  tagline: string;
  description: string;
  degree: string;
  /** @nullable */
  typicalTimeline?: string | null;
  resources: ProfessionResource[];
}