# lib/api-zod/src/generated/types/professionResource.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfessionResourceKind } from './professionResourceKind';

export interface ProfessionResource {
  label: string;
  /** Kind of resource */
  kind: ProfessionResourceKind;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  url?: string | null;
}