# lib/api-zod/src/generated/types/targetSchoolUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TargetSchoolUpdatePriority } from './targetSchoolUpdatePriority';
import type { TargetSchoolUpdateStatus } from './targetSchoolUpdateStatus';

export interface TargetSchoolUpdate {
  /** @minLength 1 */
  schoolName?: string;
  /** @minLength 1 */
  programName?: string;
  /** @minLength 1 */
  professionSlug?: string;
  /** @nullable */
  state?: string | null;
  inState?: boolean;
  status?: TargetSchoolUpdateStatus;
  priority?: TargetSchoolUpdatePriority;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  appPortal?: string | null;
  /** @nullable */
  notes?: string | null;
}